undp-transparency-portal-be/requirement.txt
django==1.11.7
lxml==4.1.1
django-cron==0.5.0
psycopg2==2.7.3.2
pycodestyle==2.3.1
djangorestframework==3.7.3
markdown==2.6.9
django-filter==1.1.0
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
geopy==1.11.0
pdfkit==0.6.1
xlrd==1.1.0
pytz==2017.3
django-redis==4.8.0
